Refactoring to allow easier access to config setting for database testing.
