+++ src/main.rs
@@ -1,6 +1,7 @@
 use std::net::TcpListener;
+use zero2prod::startup;
 
 #[tokio::main]
 async fn main() -> hyper::Result<()> {
-    zero2prod::run(TcpListener::bind("127.0.0.1:8000").expect("Failed to bind port 8000"))?.await
+    startup::run(TcpListener::bind("127.0.0.1:8000").expect("Failed to bind port 8000"))?.await
 }
