+++ src/routes/health_check.rs
@@ -1,0 +1,5 @@
+use axum::http::StatusCode;
+
+pub async fn health_check() -> StatusCode {
+    StatusCode::OK
+}
